+++ packages/core/src/js/feedback/FeedbackWidget.tsx
@@ -23,6 +23,7 @@ import { defaultConfiguration } from './defaults';
 import defaultStyles from './FeedbackWidget.styles';
 import { getTheme } from './FeedbackWidget.theme';
 import type { FeedbackGeneralConfiguration, FeedbackTextConfiguration, FeedbackWidgetProps, FeedbackWidgetState, FeedbackWidgetStyles, ImagePickerConfiguration } from './FeedbackWidget.types';
+import { hideFeedbackButton, showScreenshotButton } from './FeedbackWidgetManager';
 import { lazyLoadFeedbackIntegration } from './lazy';
 import { getCapturedScreenshot } from './ScreenshotButton';
 import { base64ToUint8Array, feedbackAlertDialog, isValidEmail  } from './utils';
@@ -335,8 +336,6 @@ export class FeedbackWidget extends React.Component<FeedbackWidgetProps, Feedbac
           )}
           {notWeb() && config.enableTakeScreenshot && !this.state.attachmentUri && (
             <TouchableOpacity style={styles.takeScreenshotButton} onPress={() => {
-              // eslint-disable-next-line @typescript-eslint/no-var-requires
-              const { hideFeedbackButton, showScreenshotButton } = require('./FeedbackWidgetManager');
               hideFeedbackButton();
               onCancel();
               showScreenshotButton();
