+++ packages/core/src/js/sdk.tsx
@@ -8,7 +8,7 @@ import {
 import * as React from 'react';
 
 import { ReactNativeClient } from './client';
-import { FeedbackWidgetProvider } from './feedback/FeedbackWidgetManager';
+import { FeedbackWidgetProvider } from './feedback/FeedbackWidgetProvider';
 import { getDevServer } from './integrations/debugsymbolicatorutils';
 import { getDefaultIntegrations } from './integrations/default';
 import type { ReactNativeClientOptions, ReactNativeOptions, ReactNativeWrapperOptions } from './options';
